# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rotaract South Asia Analytics API",
    "version": "1.1.0",
    "description": "Public, high-performance REST API and Model Context Protocol (MCP) server providing open analytics data for Rotaract and Rotary across South Asia (Zones 4, 5, 6, and 7; 44 districts; 2,870+ clubs).\n\n### Model Context Protocol (MCP) Integration\nThis service includes a Model Context Protocol (MCP) server for direct use by AI assistants (Cursor, Claude Desktop, Antigravity, VS Code):\n- **Local Stdio Command:** `npm run mcp` or `node scripts/mcp_server.js`\n- **Remote HTTP Endpoint:** `POST https://insights.rsmda.org/api/mcp` (JSON-RPC 2.0)\n- **Available AI Tools (14 tools):** `get_summary`, `get_leaderboards`, `get_districts`, `get_district_insights`, `get_zone_summary`, `search_clubs`, `get_club_profile`, `get_new_clubs`, `find_compliance_risks`, `find_dual_risk_clubs`, `get_interact_analytics`, `find_rotary_opportunities`, `get_foundation_giving`, `get_worldwide_rankings`.\n\nAll endpoints are open, unauthenticated, and CORS-enabled.",
    "contact": {
      "name": "Rotaract South Asia Analytics Team",
      "url": "https://insights.rsmda.org"
    },
    "license": {
      "name": "Open Data Commons Open Database License (ODbL)",
      "url": "https://opendatacommons.org/licenses/odbl/"
    }
  },
  "servers": [
    {
      "url": "https://insights.rsmda.org",
      "description": "Production server (Primary)"
    },
    {
...
    },
    {
...
    }
...
  "paths": {
...
    "/api/v1/clubs/new": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Newly Chartered Clubs",
        "description": "Returns newly chartered Rotaract clubs in South Asia (126 clubs chartered in the current 2026 period) with charter dates, initial members, and sponsor Rotary clubs.",
        "operationId": "getNewClubs",
        "parameters": [
          {
            "name": "district",
            "in": "query",
            "required": false,
            "description": "Filter by district number (e.g. '3000')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Filter by RI zone (e.g. 'Zone 5')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "required": false,
            "description": "Filter by base type ('Community' or 'University')",
            "schema": {
              "type": "string",
              "enum": [
                "Community",
                "University"
              ]
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort field ('charterDate', 'members', 'name')",
            "schema": {
              "type": "string",
              "default": "charterDate"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort direction ('desc' or 'asc')",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of records to return (1-100, default 25)",
            "schema": {
              "type": "integer",
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Record offset for pagination (default 0)",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Optional country filter (e.g. \"India\", \"Nepal\", \"Sri Lanka\")",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of newly chartered clubs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewClubsResponse"
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
    "/api/mcp": {
      "get": {
        "tags": [
          "Model Context Protocol (MCP)"
        ],
        "summary": "MCP Server Manifest & Tools Discovery",
        "description": "Returns the Model Context Protocol (MCP) server metadata, protocol version, capabilities, and the full schema definitions for all 14 available AI tools.",
        "operationId": "getMcpManifest",
        "responses": {
          "200": {
            "description": "MCP server manifest and tool definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "server": {
                      "type": "string",
                      "example": "rotaract-south-asia-analytics"
                    },
                    "version": {
                      "type": "string",
                      "example": "1.0.0"
                    },
                    "protocolVersion": {
                      "type": "string",
                      "example": "2024-11-05"
                    },
                    "capabilities": {
                      "type": "object"
                    },
                    "tools": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      },
      "post": {
        "tags": [
          "Model Context Protocol (MCP)"
        ],
        "summary": "Execute MCP Tool via JSON-RPC 2.0",
        "description": "Executes an AI tool via Model Context Protocol JSON-RPC 2.0 protocol (or lightweight payload: {\"tool\": \"search_clubs\", \"args\": {...}}). Supported methods: initialize, tools/list, and tools/call.",
        "operationId": "executeMcpTool",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                  },
                  "id": {
                    "type": "integer",
                    "example": 1
                  },
                  "method": {
                    "type": "string",
                    "example": "tools/call"
                  },
                  "params": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "search_clubs"
                      },
                      "arguments": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool execution result wrapped in JSON-RPC 2.0",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jsonrpc": {
                      "type": "string",
                      "example": "2.0"
                    },
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "result": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total matching records"
          },
          "limit": {
            "type": "integer",
            "description": "Records per page limit"
          },
          "offset": {
            "type": "integer",
            "description": "Record offset index"
          },
          "hasMore": {
            "type": "boolean",
            "description": "True if more records exist"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "hasMore"
        ]
      },
...
      "NewClubItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "8828727"
          },
          "name": {
            "type": "string",
            "example": "Chinnamanur Kurinji Stars"
          },
          "district": {
            "type": "string",
            "example": "3000"
          },
          "zone": {
            "type": "string",
            "example": "Zone 5"
          },
          "base": {
            "type": "string",
            "example": "Community"
          },
          "charterDate": {
            "type": "string",
            "example": "07-Sep-2026"
          },
          "members": {
            "type": "integer",
            "example": 18
          },
          "sponsorClubs": {
            "type": "string",
            "example": "Chinnamanur (Rotary)"
          }
        }
      },
      "NewClubsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewClubItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "total": {
            "type": "integer",
            "example": 126
          },
          "limit": {
            "type": "integer",
            "example": 25
          },
          "offset": {
            "type": "integer",
            "example": 0
          }
        }
      },
...
    }
  }
}